fix: changed offset type to long

+++ virtual-pipeline/src/main/java/io/jenkins/plugins/VirtualPipeline/VirtualPipelineLineOutput.java
@@ -12,17 +12,17 @@ public class VirtualPipelineLineOutput {
 
     private LineType type;
 
-    private int lineStartOffset;
+    private long lineStartOffset;
 
     public LineType getType() {
         return type;
     }
 
-    public int getLineStartOffset() {
+    public long getLineStartOffset() {
         return lineStartOffset;
     }
 
-    public VirtualPipelineLineOutput(String regex, String line, int index, Boolean deleteMark, LineType type, int lineStartOffset) {
+    public VirtualPipelineLineOutput(String regex, String line, int index, Boolean deleteMark, LineType type, long lineStartOffset) {
         this.regex = regex;
         this.line = line;
         this.index = index;
